backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10          # sync driver for alembic + seeding

# ── Config ────────────────────────────────────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# ── HTTP / scraping ───────────────────────────────────────────────────────────
httpx==0.28.1
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.3.0
playwright==1.49.0               # for JS-rendered portals

# ── Anthropic ─────────────────────────────────────────────────────────────────
anthropic==0.40.0

# ── ML / data ─────────────────────────────────────────────────────────────────
xgboost==2.1.3
lightgbm==4.5.0
scikit-learn==1.6.0
imbalanced-learn==0.12.4
pandas==2.2.3
numpy==2.2.0
joblib==1.4.2
scipy==1.14.1

# ── NLP ───────────────────────────────────────────────────────────────────────
spacy==3.8.3
transformers==4.47.0
torch==2.5.1                     # CPU build; GPU: torch==2.5.1+cu124

# ── Task queue ────────────────────────────────────────────────────────────────
celery[redis]==5.4.0
redis==5.2.1
kombu==5.4.2

# ── Notifications ─────────────────────────────────────────────────────────────
sendgrid==6.11.0
slack-sdk==3.34.0

# ── Utilities ─────────────────────────────────────────────────────────────────
arrow==1.3.0
tenacity==9.0.0
structlog==24.4.0
sentry-sdk[fastapi]==2.19.2
mlflow==2.19.0

# ── Auth & Security ───────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# ── Rate limiting ─────────────────────────────────────────────────────────────
slowapi==0.1.9

# ── Caching ───────────────────────────────────────────────────────────────────
redis[hiredis]==5.2.1
cashews[redis]==7.3.0

# ── Fuzzy matching (entity resolution) ───────────────────────────────────────
rapidfuzz==3.10.1

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
pytest-mock==3.14.0
factory-boy==3.3.1
